requirements.txt
# Core dependencies
pandas>=1.4.0
numpy>=1.22.0
scipy>=1.8.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.9.0

# Deep learning
torch>=1.12.0
torch-geometric>=2.0.0
torchmetrics>=0.9.0

# Graph and network analysis  
networkx>=2.8.0
snf>=0.2.2

# Dimensionality reduction and clustering
umap-learn>=0.5.0

# Survival analysis
lifelines>=0.27.0

# Experiment tracking
wandb>=0.13.0

# Progress tracking
tqdm>=4.64.0

# Development and notebooks
jupyter>=1.0.0
notebook>=6.0.0

# Optional: For advanced methylation preprocessing
# methylcheck>=1.0.0  

# Optional: For interactive data exploration  
# bamboolib>=1.30.0